Add files via upload
I went through all the scripts and made some modifications. In this version, you can use python total_control.py <0|1> to prepare your system. An example of the output can be found in the system1 folder. This file is ziped and all the trajectory files in pre_equ foler are missing because there is a limit on the size of uploaded files.